added test

+++ pandas/core/array_algos/quantile.py
@@ -184,5 +184,5 @@ def _quantile_ea_fallback(
     try:
         out = type(values)._from_sequence(res, dtype=values.dtype)
     except TypeError:
-        out = np.atleast_2d(np.asarray(res))
+        out = np.atleast_2d(np.asarray(res, dtype=np.float64))
     return out
